requirements.txt
patchify
torch
torchvision
torchmetrics
pillow
tqdm
matplotlib
pandas==1.5.3
opencv-python